Create keyscan.c
Some cheap keyboards are not connected to a supporting hardware device and is a bit harder to read. The ML2 Keyboard can act in this simple mode if choose ML5 mode.
The principle for this mode is described below.
The keyboard is built up by four row connectors and four column connectors. At the 16 crosses between raw and column there is normally no electrical connection. At each cross there is a key and if key is pressed there will be an electrical connection between the column and raw.
All columns are on one side connected to a voltage source (5V) through a resistor and the other side of the columns are connected to the four least significant bits (b3 - b0) on a 8 bit out port connections from the keyboard.
All four rows are through an inverter connected to the least significant bits (b3 – b0) on a 8 bit in port connection to the keyboard
The keyboard is connected to the computer system by two 8 bit I/O ports. The system INPORT ( Address 0x09C1) reads the column value through bit 0 to bit 3 and the system OUTPORT (0x09C0) controls the row values by bit 0 to 3. All as the figure above shows.
If no key is pressed the value on INPORT[3..1] (columns) always will be high on all four bits. If all four bits on OUTPORT[0…3] are zero the value on all rows are high due to the inverters. If any key now is pressed it will not give any change on any BITS on INPORT (column bits). They will all still be high.
The only way to detect a pressed key is to set a row to low (0) by setting the corresponding row bit in OUTPORT to high (1). If any key in that row now is pressed it will give the corresponding column a zero value that can be detected by reading the INPORT.
To read the keyboard you must in a loop clear one rows value (set the corresponding OUPORT bit) and check if INPORT is not equal to four set bits. If still four set bits you clear next row and check the columns value. As soon as you find a column value with not four set bits you have detect a pressed key. If so you know which row and column and you can from that calculate which key number is pressed. The pattern for these two nibbles (nibble = four bits) is unique for each key. If you check all four rows without finding any INPORT value not equal to four set bits there is no key pressed for the moment.
We can assume that Key is numbered as from top row from left 0 , 1 , 2 , 3 , 4 and next row 5 , 6 , 7 , 8 etc.